feat(bot): rework on structure
